+++ src/components/settings/userMgmt/MagicStringModal.tsx
@@ -7,41 +7,10 @@ import { showToast } from '../../../utils/ui/toast';
 import {
   copyTextToClipboard,
   encodeServerMagicString,
+  magicPayloadAddressFields,
 } from '../../../utils/server/serverMagicString';
 import { shortHostFromServerUrl } from '../../../utils/server/serverDisplayName';
 import { useAuthStore } from '../../../store/authStore';
-import {
-  normalizeServerBaseUrl,
-  serverShareBaseUrl,
-} from '../../../utils/server/serverEndpoint';
-import type { ServerProfile } from '../../../store/authStoreTypes';
-
-/**
- * Pick out the dual-address fields for the profile that owns `serverUrl`,
- * so the encoded invite carries the share URL + alternate + share flag
- * when those exist. Falls back to the raw `serverUrl` for legacy / single-
- * address profiles — v1 wire shape unchanged.
- */
-function magicPayloadAddressFields(serverUrl: string): {
-  url: string;
-  alternateUrl?: string;
-  shareUsesLocalUrl?: boolean;
-} {
-  const normalized = normalizeServerBaseUrl(serverUrl);
-  const match = useAuthStore.getState().servers.find(
-    (s: ServerProfile) =>
-      normalizeServerBaseUrl(s.url) === normalized ||
-      (s.alternateUrl != null && normalizeServerBaseUrl(s.alternateUrl) === normalized),
-  );
-  if (!match) {
-    return { url: serverUrl };
-  }
-  return {
-    url: serverShareBaseUrl(match),
-    ...(match.alternateUrl ? { alternateUrl: match.alternateUrl } : {}),
-    ...(match.shareUsesLocalUrl ? { shareUsesLocalUrl: true } : {}),
-  };
-}
 
 interface Props {
   user: NdUser;
@@ -96,7 +65,10 @@ export function MagicStringModal({
         return;
       }
       setSubmitting(false);
-      const addressFields = magicPayloadAddressFields(serverUrl);
+      const addressFields = magicPayloadAddressFields(
+        serverUrl,
+        useAuthStore.getState().servers,
+      );
       const str = encodeServerMagicString({
         ...addressFields,
         username: user.userName,
